Access SchemaStore with correct KernelVersion
With the introduction of token indexes there is a intermediary state when on 4.3 binaries before
upgrade has been run that it is necessary to inject a schema rule that represent the former
label scan store. Now only the methods where you have to make an explicit choice with which
version to open the store are kept.