Checking cluster config in Single Instance
We may enabled the discovery service in Single Instance all config checks were bound to CORE and READ_REPLICA.
Now we also need to check if SINGLE + clustering enabled.
In case of SINGLE if discovery type is LIST (this is default) the 'initial discovery members' need to contain the instance's own
advertised discovery address (otherwise topology service won't find itself and start properly). In a case of CORE members this is
also suggested, but not mandatory - so we just warn in log.
We allow 'initial discovery members' to left empty in SINGLE (as this the default), so the user only needs to flip the switch for
enabling clustering. If 'initial discovery members' is empty the machine's advertised discovery address will be used.