ConsistencyChecker relationship group race condition fix
Fixes an race condition in the RelationshipGroupChecker, where two different threads would race write to the same node in the cache. This would lead to an incorrect report of an inconsistency, as the cache got incorrect.

cl: Fixes an issue where the consistency checker would incorrectly report "The first relationship group record has another node set as owner." inconsistencies.