OrLeafPlanner: no coveringPredicates
If all branches of an OR would have a PartialPredicate with the same coveringPredicate
as solved, the OrLeafPlanner would mark the coveringPredicate as solved.

This is wrong, but the bug did not surface because we usually do not get plans
with PartialPredicates at this point.

But, in theory, it could lead to such a situation:

```
.distinct                 // solves n.prop > 5 (This is wrong.)
.union
.|.indexScan("n:M(prop)") // solves PartialPredicate("n.prop IS NOT NULL", "n.prop > 5")
.indexScan("n:N(prop)")   // solves PartialPredicate("n.prop IS NOT NULL", "n.prop > 5")
```