Fix SlottedPipeOperator to use correct parallel query state
We cannot get the feed-pipe query state up-front when scheduling and
cache in the task, since the QueryResources/WorkerId may not be the same
as when we run the task.
Instead we need to do it in every operate call so that we can correctly
handle continuations in parallel runtime.

- Make SlottedPipeMiddleOperator essentially a stateless operator
- Adapt SlottedPipeFallbackTestBase to use NonPipelined to test
profiling of slotted pipe middle operators