Tests for consistency checker with injected NLI
Starting an existing database on 4.3 or newer binaries with token index feature on
makes the old label scan store into a token index. Until upgrade transaction has
been run there will be no schema rule in the schema store for this NLI, just
a rule we inject in SchemaStorage. This tests that consistency checker can check
the NLI with this injected rule.