Removes accidental relationship type ID limitation in importer
When a node becomes dense during import the relationship groups
are kept in memory in a chain, one for each group.
The group "field" in this cache was only 2 bytes, which would
not fit the whole relationship type ID range of a high_limit store.

This relationship type limit is now increased to 3 bytes to
be able to hold type IDs that are much larger. The memory overhead
of this is of course 1 byte per dense node and type during import,
or put differently it increased by ~2% for dense nodes, which is
not very much.