Adjust test names to reflect their assertions.
The assertions in these two tests have changed in
https://github.com/neo-technology/neo4j/pull/2324

This change allowed different cached properties that are use
in Union branches to survive past the Union.
The reason was that there is little overhead in writing to cache,
and it can give benefits for those rows where we can read from the cache.

It was missed to update the test descriptions back then.