Make sure periodic commit works with batching
Adjust the batch size used by pipelined runtime to the explicit batch size
given in the periodic commit clause. If the explicit batch size is big we
try to find a denominator within a reasonable range, or something close to
a denominator if one cannot be found (i.e. in this case we allow some drift
between explicit batch size and commits).