Fixes an issue with RecordNodeCursor#hasLabel
Sometimes, for a larger set of labels, i.e. when dynamic label records
were used, the HasLabelSubscriber (used to implement this method)
could sometimes think it had labels that it didn't have.
The issue was that HasLabelSubscriber didn't account for the
"number of bits in last byte" information from the dynamic array header

Fixes #12704