Convert page cache feature toggle into new settings `dbms.store.files.preallocate` to be able to control store file pre-allocation by page cache.
On some file systems stores with enabled preallocation can consume significantly more space comparing
to stores that were created by the database with a disabled option. For example, on BTRFS on some setups,
we can consume up to  30% more space for the store(allocated but not used).
While space can be reclaimed and fs can be rebalanced, for each TB we are losing around 300GB in those scenarios.
To be able to cover those such use caces we will allow preallocation to be configurable instead.