neo4j-admin console create pid file on unix
We try to write the neo4j.pid file. If we fail because of permissions issues we log the error and start anyway. This is to not break behavior with 4.3 where write permissions in the run directory can not be guaranteed. This allows `neo4j-admin report` to find running neo4j instances also when started with console. We only do this on unix, as windows does not deal with pid files

cl: neo4j-admin console create pid file on unix