Explicit write cursor propagation
Before this PR every record update was creating and closing new page cache cursor.
That implies that every time we need to create a cursor, lookup page, lock it, position cursor, etc
In reality, a lot of writes and updates are not that independent and they always come together.
To be able to control scopes of those write cursors this PR changes updateRecord calls to always take
corresponding store cursor and set of other cursors for any dependent updates.
The scope of that cursor is controlled by a caller.