Always expect the caller to provide record read cursor
Previously record store had a default method that
was opening and closing read cursor that was used for
single record reading. This PR removes that method completely and
demands from caller to provide cursors for callee instead.
Some places were updated to open and close cursors in place
to simulate current behavior. That will be changed in subsequent PRs