Use BeforeAll / AfterAll in KernelAPIReadTestBase
@BeforeEach was used to setup the database once, and to reuse it each
time, as testSupport is only null on the first test.

Likewise the null check can be removed in the @AfterAll, testSupport
will never be null at that point.

Thus the graph setup was moved to @BeforeAll, leaving the switching of
user credentials in @BeforeEach.

The TestInstance was already set with LifeCycle.PER_CLASS allowing
non-static @{Before,After}All.