Don't print queries in debug.log
Whenever queries were discarded from the cache we printed a message looking something like

```
Discarded stale query from the query cache after 10 seconds. Query: RETURN 'super secret internal message'
```

This poses a security problem especially in AURA where we are not allowed to see `query.log` but can see `debug.log`.
By just waiting for queries to be discarded from the cache we can in principle deduct all the queries run on the database.

This PR changes so that instead of writing the full query we give the query id so that you can go into `query.log` - if available -
to see what query that was discarded.

changelog: Don't print queries in debug.log when they are discarded from the query cache.