Let PartialSort decrease effective cardinality of children.
PartialSort always has to Sort whole buckets, even when under a Limit.
The work reduction from the Limit does therefore not propagate 100 % to the
child of the PartialSort.

We have no way of determining the size of buckets currently, so we simply guess
that the work of the child is 10 % more than the work of the PartialSort.

Co-authored-by: Arne Fischereit <arne.fischereit@neo4j.com>