Don't overflow for -Long.MIN_VALUE
Currently this will happen:

```
neo4j@neo4j> return --9223372036854775808;
+-----------------------+
| --9223372036854775808 |
+-----------------------+
| -9223372036854775808  |
+-----------------------+
```

That is we silently overflow when trying to create the number `9223372036854775808`,
which is the same as `Long.MAX_VALUE + 1` and is beyond what we can handle in the database.
This PR changes it to throw the same type of error that would of happened if you try to
do `RETURN 9223372036854775808` directly.

fixes #12726
changelog: Don't overflow on `--9223372036854775808`. Github issue: neo4j/neo4j