Detection of property chain cycle on read
RecordPropertyCursor and dynamic value record reading will now be aware of
chain cycles and be able to detect and abort reading if getting stuck in such.
The detection is cheap, just a counter up until it reaches 100 000, at which
point a set is instantiated and duplicates detected from that point on.

StoragePropertyCursor got an additional init(long,long)
method where the second long is the owner reference to include
in an eventual exception if hitting the cycle detection threshold.