Block creation of index backed constraint that is similar to constraint dropped in same tx ()
* Block creation of index backed constraint that is similar to constraint dropped in same tx

This is the result of a long investigation on all issues with dropping
and creating equivalent index backed constraints in the same transaction.

ConstraintDescriptor equals method does not take id and name into
account which means the constraint drop and constraint create
are mixed together in transaction state and instead of creating a new
constraint, the drop of the old would simply be removed from
tx state and this would cause a very weird state where new
backing index would never come online and the old constraint and
the new index would not be linked together, schema store would be
inconsistent because of the mis-linked schema rules and subsequent
index updates would fail with IllegalStateException.

Even if we fix the equals method, this would still be problematic
because if we crash while populating the new backing index we
will end up with two indexes on the same schema. This is something
that we don't allow currently.