Use faster index checking for node indexes in experimental consistency checker
There was a bug in the experimental consistency checker that made it seem like no
indexes had the value capability (the actual values stored in the index).
This meant that indexes were checked one entry at a time while consistency checking
the node store. Now they use the nice IndexChecker as intended.
The IndexChecker goes through the index entries sequentially and should be give a
nice performance improvement.

Co-authored-by: Petr Janouch <petr.janouch@neotechnology.com>