Fix leak of execution plan into data collector buffer
The data collector RecentQueryBuffer can hold up to 8192 queries.
A buffer entry held a reference to an ExecutionPlan in a lambda,
in order to generate a plan description lazily if needed.
This would prevent code generated classes from being garbage collected.

This commit isolates from the execution plan just the parts needed to build
the plan description instead.