Don't fold sensitive literals
When we do fold constants we rewrite things like `2 + 3` to `5` but if obfuscate literals
is enabled this breaks down because we 1) the new literal `5` wasn't created as a sensitive literal
and 2) in the query log we write the unrewritten query so we need to preserve the positions of the two
sensitive literals. This PR changes so that if obfuscate literals is enabled we don't do any constant folding.