Fail on `MERGE ({p: NaN})`
Since `NaN != NaN` this MERGE will never MATCH and always create a node or
relationship, which breaks the assumption that multiple identical merges
should only create a node or relationship the first time we run the query.
In this PR we disallow `NaN` properties in `MERGE` in the same way as we
do for `null` properties.