Fabric shutdown transaction timeout
Before this PR on the shutdown of DBMS fabric TransactionManager was unconditionally terminating all running transactions. That is not entirely correct since it was breaking the sequence of components shutdown and was ignoring at least transaction shutdown timeout.
This PR update TransactionManager to terminate only transactions that are non-local (one that runs on some remote envs) and will allow DB local transaction to continue its predetermine sequence and will let owning databases deal with them.