Making count stores more robust
Before this change, a count store would panic when instructed to store a negative count,
which would in turn panic the entire database.

Even though negative counts should not happen when everything works as expected,
they can be a side-effect of a corruption. For instance, if a new type of relationship
materialises in the stores as a result of a corruptionand it will not be accounted for
by the count store. If the relationship is later deleted, the count for this type of record
might go negative as unlike the creation, the removal will be accounted for.

The main purpose of this change is to prevent the database from panicking in such scenarios
and try to carry on even with some degenerated functionality.

When a negative count is observed, information that the count for the given key is invalid
will be store in the store and subsequenlty an exception will be thrown when someone tries
to access the count for that key. Also since the value for such key cannot be trusted anymore,
its updates will be ignored.
This is not ideal, but still better than panicking the database. It is acutally not so bad as
it might seem, because it is very unlikely that someone will ask about a count of something
created by corruption.