Support larger relationship type in relationship group commands
There was a bug in the high-limit format that truncated the relationship type in the relationship
group command in the transaction log. This could result in a corrupted store during recovery for
a single instance, or on followers in a cluster.

This introduces a new pseudo-command type that is used when the type id overflows a short.

Co-authored-by: Anton Klarén <anton.klaren@neotechnology.com>