4.4 reintroduce bolt tx manager ()
Introduce Transaction Manager into Bolt Server

This change is the first part of introducing a Transaction Manager interface which will ultimately unify the way Bolt Server and HTTP servers interact with Kernel.

The new interface TransactionManager provides a central point to manage and execute transactions. The first implementation of this interface is the StatementProcessorTxManager which utilizes the existing transaction manager implementation used by Bolt Server.

In part two the TransactionManager will be introduced into the HTTP server and will enable HTTP requests to be routed when server-side routing is enabled.