Add PartitionedScan variant of node property index scan
Using the example of the current, non parallel, void returning scan;
only the existence first property of the index was queried.

This fails on composite indexes; therefore, testing the existence of all
properties in the index allows for composite indexes to be scannable.

A quick test indicates that the non parallel, void returning scan does
not need this. Will have to investigate why the partitioned scan does.