Reserve bytes in the pages header
Allow pages to have configured number of reserved bytes in the head of the page.
Old formats have 0 reserved bytes and there is no way of how that can be changed.
Setting change for an existing store is unsupported.
Migration of this setting is unsupported.
By default, we keep 0 bytes as reserved in the latest record format.