Set checkpoint log version field on recovery
Recovery is an indication that the store was not cleanly shut down and a sign that the metadata store can potentially contain untrusted data. And while the recovery process was using only available checkpoint files as a source of information about data sources it did not update the metadata store field with available checkpoint log version information on recovery completion. That could result in a wrong value in the metadata store field that will be used during the creation of a subsequent checkpoint file. And that can cause problems on subsequent restarts since the sequence of checkpoint files will be broken.
This PR fixes that and updates the checkpoint log version field as result of recovery, as we did before for tx log version files.