Fix issue where connected relationships were added separately as `PatternExpressions`.
In `OptionalMatchRemover`, if two connected relationships can theoretically be extracted, like in

```
OPTIONAL MATCH (a)-[r:R]->(b)-[r2:R2]->(c)
RETURN COUNT(DISTINCT a) as count
```

then both relationships would need to be part of one `PatternExpression`.
But they were added as two separate `PatternExpressions`.
This changes the behavior such that overlapping relationships are not extracted,
unless the overlap is on an element to keep anyway, like in

```
OPTIONAL MATCH (a)-[r:R]->(b), (a)-[r2:R2]->(c) WHERE b:B AND c:C
RETURN COUNT(DISTINCT a) as count
```