Complain about out of disk space more open and aggressive
Previously failures to preallocate files transaction log or store file would be not distinguishable between each other in was just log warns for transaction log cases.
This PR introduces a way how we can find out that failure is actually because of OOD and complain about that on error level or throw exceptions with more verbose messages.
This is an intermediate state of things since ideally, we want to finish in a state where the database will mark itself as read-only.