Pave the way for tracking memory of different transactions at runtime.
A QueryMemoryTracker is not bound to a transaction any more,
and only captures the high water mark for the whole query, possibly spanning multiple transactions.

A QueryMemoryTracker can hand out a MemoryTrackerForOperatorProvider, which can be used to get
a MemoryTracker for one specific operator. This MemoryTracker is transaction bound.
It forwards all its calls to both the transaction memory tracker and a
query global operator memory tracker. This query global operator memory tracker is used to obtain
the correct high water mark for each operator, possibly spanning multiple transactions.

Setting a single memoryTrackerForOperatorProvider in `StateFactory` is not ok for now,
but needs to change when pipelined should support CALL TRANSACTION.