Remove support for rewriting equality predicates in PatternExpressions
Property equality predicates were previously extracted into `PatternExpressions` if possible.
Unfortunately, relationship property equality predicates on relationships to extract were
simply discarded instead of put into the `PatternExpression`.

Moreover, this code was only exercised in the unit tests. When invoked with the rest of the
pipeline, equality predicates had already been rewritten to `In`, making it practically dead code.

Instead of fixing this to rewrite `In` and fixing it such that relationship property equality
predicates are correctly extracted, this maintains the current behavior and removes the
effectively dead code.