Lightweight transaction stamp to track open transactions in id reuse conditions
It was observed that accumulating transaction snapshots over time keeps way too many
things in the heap. Over time transactional snapshot swelled and now contains lots of references to a lot of things that we do not need if we only want to know if the transaction is closed now. As result, this PR introduces KernelTransactionStamp that keeps reference only to transaction and reuse counter.