Remove Exists predicates from Fusion capability
The Exists predicate is not wholey within the BTree part of the Fusion
index; and thus if supported by current means, only the non Text valued
properties would be returned.

This also prevents being able to scan over the index in this way, as it
too would depend on bridging the underlying indexes. If there is a
simple way to partition a Lucene based index, this can be revisited.