Leaving transactions in implicit context as implicit
In cases when we were to start an implicit transaction, but the query would not
start with "PERIODIC COMMIT", we would not start an implicit but rather an
explicit transaction. This needs to change, as we also only want to allow call
transactions to take place in implicit transactions and the clue cannot be
determined by the pre-parser anymore.

Exceptions are then thrown from Neo4jTransactionalContext#contextWithNewTransaction

Co-authored-by: Satia Herfert <satiaherfert@gmx.de>