Volume-based checkpoint threshold
Introduce a checkpoint policy that is based purely on the amount of transaction logs written. Idea is to use the volume of the written tx log as an indication of how much data the database needs to replay during recovery. The amount of data that should be replayed should have quite a good correlation with the amount of time the database will need to spend recovering those transactions.