Less surprising use of TokenIndexEntryUpdate internal arrays
Previously the before/after arrays (containing token IDs) were,
upon being fed to a TokenIndexUpdater, changed in-place to
represent removals/additions. This reuse of what was expected to be
final and immutable arrays was surprising and has caused some
headaches in development and debugging.

Now instead the TokenIndexEntryUpdates that come in gets converted
into LogicalTokenUpdates which contains clones of before/after.
This entails more allocation, but will hopefully not
degrade performance.