Fix nested plan predicate in optional expand
Slot allocation assumes that the predicate will be evaluated on the incoming
row but for OptionalExpand this is not true so queries with NestedPlanExpression
in the predicate will fail in physical planning.

In released version we fix this by forbidding rewriting to OptionalExpand whenever
there is a NestedPlanExpression in the predicate, however in `dev` we should attempt
to fix the actual slot allocation problem.

fixes #12757

changelog: Fix error from query using OPTIONAL MATCH and predicate with a pattern expression, Github issue: neo4j/neo4j