Introduce transaction log flush counter metric.
Introduce a new metric to represent the number of transaction logs flushes.
Flush is the most expensive part of the transaction log-append process. With a newly introduced setting that can affect log buffer, there is a way how users can influence that by allocating additional memory to transaction logs buffer.