Make TokenPredicate implement IndexQuery
IndexQuery originally was implemented as a blank interface to link the
conceptually similar PropertyIndexQuery and TokenPredicate together, in
and be accessible in the internal.schema package.

As this was for the benifit of IndexCapability::supportPartitionedScan,
and TokenPredicates always passed; it was "simplied" to just have
something in internal.schema that could cast to PropertyIndexQuery.
I moved a few things that could be useful (and in scope) from the
abstract PropertyIndexQuery to the IndexQuery interface, so it wasn't
empty. These methods were not used from IndexQuery, outside of
IndexCapability::supportPartitionedScan.

All this was squashed and commited as 5c4c8ea2e8c.

I will be reverting some of those decisions, to ensure changes that
changes that influence assumed invariants, are caught.