Wrong result in pipelined runtime
Whenever we evaluated an expression containing cached properties in the leaf of
the RHS of an apply and if that leaf operator didn't advance the cursor we got
stale cached properties leading to the wrong result. For example

```
MATCH (a:A)
OPTIONAL MATCH (b:B) WHERE a.id = b.id
WITH a.id AS id, b
RETURN b
```

if there is an index on :B(id).

changelog: Fix wrong result in PIPELINED runtime for some cases of OPTIONAL MATCH