Ability to collect page cache metrics on a file basis
Enhance existing page cache tracing infrastructure with
page file swapper tracers to be able to trace and associate
trace events with individual page swappers, which are individual
mapped files from the user's perspective.
As a result of this PR, we have a possibility to observe page cache metrics for
individual files to be able to see individual store/indexes access patterns,
popularity, demands, etc. This should give us data about future PC quotas,
different priorities, reservations, etc.

Please note that as a result of this PR new set of metrics is not exported by default.