Fix kernel level range queries on arrays of points
The points within PointArrays were stored in the index using the space
filling curve; however, when queries for the range, this was not taken
into account. This resulted in false positives being returned, as no
extra filtering was done.

To avoid the complexity of frequently query the index multiple times for
each point within the array; the simpler option of simply filtering over
the whole range was prefered.