Always include group degrees store in file listing
Because e.g. consistency checker expects it to be there.
Not including it in file listing for a store that hasn't yet
been "dbms.upgrade"'d to V4_3_D4 was a remnant from not creating
it until it had seen that new version. Now it's always created
and is expected to be there.