Update VmPauseMonitor defaults to match G1 defaults
We ship with default G1 settings out of the box.
And default target value for `MaxGCPauseMillis` is 200 ms so it's strange to warn about any pause that is more than 100ms if we are fine with target 200 ms pauses.
This PR changes default timeouts for VmPauseMonitor to be 200 ms instead.