Make page cache number of pages to keep free a percentage of the size
Before this PR number of pages that PC should keep free for new data was only 30 pages. That number was quite a law and in the majority of cases, we fell back into cooperative eviction by transaction executors.
This PR changes that limit to be 5% of page cache capacity by default.
We will still have an absolute limit of 30 pages as a minimal number of pages to keep free that we can ever have. And limit can be bumped as high as 50% of the total capacity.

Measurements were performed for page caches with different sizes: 8m, 128m, 1024m, and different numbers of free pages: 30, 5%, 15%, 25%. And the percentage of pages is always faster with not that big difference when the number is increased. The difference is observed much better if we eliminate tx appending(no logs and writes to the store only) in test workloads.