Handle concurrent stop and close in ForsetiClient
Previously apparently when concurrent calls to ForsetiClient will try to stop and close on it
one of the threads would most probably fail with internal exceptions that were related to lock client state cleaning.

This PR change behavior of stop and will prevent that case to happen by not allowing stop to be called when the close was called.
Also close will now wait for a stop to complete before proceeding.
Please note that from now on a subsequent call to stop we throw LockClientStoppedException