Early eviction of closed files pages
Before this PR pages of closed paged files were not evicted on close,
unless we were running out of swapper ids, which is not such a common event.

First of all, it was hiding the case when eviction actually happened, but cursors
were not ready for not present page, as result additional check on `shouldRetry`
for the already closed file was added.
This PR changes that behaviour and now we will try to evict pages of the closed files
when we do actually close them. closing is not a hot path, and we should not put
eviction of pages into a readers bucket.

We still have a case when we can't evict all pages on close of the file (cursors are still open for that file),
those will be processed in a delayed manner. While it's quite an uncommon case,
PR still changes the threshold for that and made sweep more proactive to have
as many pages that can be used by readers as possible.