Fix re-creation of expression cursors on periodic commit
After a periodic commit we need to recreate ExpressionCursors with the
CursorFactory and CursorContext from the new transaction, so that they
do not remain to be reused in a closed state while they have also been
returned to the cursor pool in the closed transaction, which can later
become reused for another transaction.
This problem also occurred with the cursors used to populate the result
in ProduceResult in slotted and interpreted runtimes.

- Use ExpressionCursors in ProduceResultsSlottedPipe and ProduceResultsPipe, so that they also get properly recreated.