Make ForsetiClient respect equals hashCode contract.
Before this PR ForsetiClient after each initialize call would change hashCode and equality fields values
That will lead to cases where same client will not be found in a set's that used over the lock manager.
And in scenarios like:
1. set.add(client)
2. client.initialize
3. set.remove(client) <- nice try!
that will lead to absolute failure to do somewhat desired.