Split QueryContext into read-only and read-write parts
This is mainly to make it more explicit what components actually need
write capabilities, in preparation for supporting parallel thread
execution contexts for read-only queries.

Also try to hide exposure of some fat interfaces from QueryTransactionalContext
and use explicit wrapper method for exposing the things that we actually
need. This is to make it easier to maintain having a read-only parallel
transactional context. More work remains on this.

Cleanup ShortestPath that does not need transaction anymore.