Disable io limiters only on global shutdown
Before this PR database stop and recovery were not respecting configured io limits
and always flushed data as fast as they can. This PR changes that and respect the
limits when individual db stopped or recovered.
The only time when the limit is disabled for all databases - during dbms shutdown
since we want to perform shutdown as fast as possible and there is no transactional workload anymore.