Fix failing tests
When profiling is enabled, we need to have a profiling tracer set on the
cursors used to populate results in ProduceResult pipes to get correct
db hits count. Since the db hits counting is done a bit differently in
interpreted and slotted we cannot do this on the ExpressionCursors
in the QueryState. Instead we create an ExpressionCursors copy through
the QueryContext for the ProduceResult pipe, and override the create
method in ProfilingPipeQueryContext to set a tracer.

(cherry picked from commit b377341816a245264485056d6ac8a2affac13c39)