Fix ALL_ENTRIES scan with TxState
Query the index capability for what values it can provide and store;
and filter the transaction state updates accordingly.

The fusion index previously did not override completeConfiguration,
and thus just passed the IndexDescriptor through without change.
This kept the IndexCapability to the default NO_CAPABILITY, which
therefore indicated that the index could not accept values of any
catagory. Adding a LuceneIndexCapability, was needed to resolve this
issue; and thus now correctly indicates that it accepts TEXT.

Add IndexCapability to TestIndexDescriptors