Removes unnecessary dependencies on record-storage-engine
Now there are no unnecessary compile dependencies on
record-storage-engine, except for test modules
(which is a later problem to deal with).

neo4j/neo4j-enterprise only has runtime dependency because
it needs to include it in packaging.

Also batch-insert, import-tool and consistency-check modules
still depends directly on record-storage-engine because they
are tools specific to that storage.

And then there's the tools module which is a jumble of
small tools that sometimes use specifics of this storage,
but that's also a smaller thing to deal with later.