Moving ReadAheadChannel close check
Before this change, the check if the channel is closed
was performed with every byte read. The check is not expensive,
but it is not cheap either given the frequency of invocation.
The reason why it is not super cheap is that the channel
usually represents a pipeline of channels and the check
is propagated through it.