Introducing a new way of loading pages during cache re-heating.
The old algorithm goes through each file one by one and pages
in a file are loaded using very aggressive parallelism if it seems
like they reside on 'high IO device' ( whatever the term
is supposed to mean ).
The aggressive per-file parallelism turns what is mostly sequential
IO into random one with negative consequences on most IO systems.

The new algorithm processes files in parallel, but each file is
processed sequentialy. This way, clusters of neighbouring pages
are loaded more efficiently as most OSes and IO systems detect
sequential access and do something smart.

For instance with a databse deployed on EC2 EBS, this resulted
in doubling the performance.
With a databse deployed on a directly attached NVMe, this resulted
in 30% improvement.
It is also expected that the old algorithm influences more negativelly
concurrenlty running workloads especially in IOPS limited environment,
but this is harder to measure.

Lastly, the heuristics for detecting 'high IO device' are a bit weird
as they identify EC2 EBS with 5000 IOPS as a 'high IO device'.

The old algorithm has been moved to its own class and can be selected
using an unsupported setting for the unlikely case that there are
some environemts where it performs better.