Fix Core API planning of findNodes and findRelationships together with indexes ()
findNodes, findNodes, findRelationship and findRelationships now correctly plan against indexes based on what the indexes has support for. Most important commits in this squash:

* IndexCapability can test query support for token lookup queries
Add value category to TokenPredicate. Update javadoc for IndexCapability.supportPartitionedScan to not reference things that cannot be reached.

* All index readers correctly call IndexMonitor.queried( IndexDescriptor descriptor )

* Transaction.findNode and findRelationship now plan correctly against indexes

* Include FULLTEXT index when testing findNode and findRelationship

* Properly test that we find the entity

This revealed an issue with string search:
When cutting a random string up using String.substring,
some exotic symbols would be cut in the middle and would
not be included in the result set. To work around this
we only use symbols from the Basic Multilingual Plane.

* @RandomSupport.Seed annotation can be used on class level if lifecycle is PER_CLASS