Point in time (transactional history) recovery
Before this PR recovery was always replaying all the changes from transaction logs and there was no way for external users to influence that behavior.
This PR introduces the possibility to finish recovery earlier and that is controlled by user-provided criteria.
Newly introduced supported criteria's are transaction id and transaction date based.
Using those users will have the ability to stop recovery when some transaction id or date is reached.

The new way of recovery is supported only with new differential backups. All previous usages of recovery are working as before and always replay all transactions.
Criteria for early recovery termination can be passed by using `--restore-until` option.
Supported values for the option are:
      * long value - criteria based on transaction id, will replay transactions up to the provided value (excluding);
      * date-time - criteria based on the transaction date, will replay transactions with a date up to the provided value (excluding).
Point in time recovery is only possible for transactions that are not covered by any checkpoint yet.