Add end offset to LogChannel
We will need this in the catchup service in order to know where to stop streaming the log file.
I'm not sure if using the last closed transaction is a good idea, but it's seems to be the most simple solution right now (as it is the only position kept by the metadata store), and I can't see how it could cause any issues.