Fix cursor close issue with expand into in pipelined runtime
Let the ExpandIntoOperator manage the traversal cursors used
with CachingExpandInto so that it can use the pipelined runtime cursor pool.

Fixes an issue where the traversal cursor would be closed and returned
to the transaction cursor pool from CachingExpandInto, while also being
freed for reuse to the runtime cursor pool in a closed state.

Cherry-picks: