Revert "Write checkpoint log entries with current kernel version"
This reverts commit 3397dece
Trying to write checkpoints with the current kernel version instead of always
using the latest version was a problem for a rolling upgrade scenario when
upgrading from 4.2. In 4.2 we didn't have a kernel version in the metadata
store. When copying a store from a 4.2 member to the member started with the
newer binaries we would during recovery of that store try to write a checkpoint
and fail when trying to get the kernel version.
Writing with the latest version is not a real problem since the checkpoints
are never sent around in the cluster, so we revert this for now and introduce it
again in 5.0 where it won't be a problem since there will be an offline migration to
5.0.