Fix Fulltext index updates involving unsupported value types
Previous IndexUpdateIgnoreStrategy was a quick patch to fix the
apparent issues at the time. The bug this set of commits fixes, was
found by reexamining that patch.

Once the actual bug was found, this required a different patch, and
relocating of the ignore check, to comply with the other changes; and
still have the semantic check.

Most changes are in the shared code for the other Lucene based indexes.