Add start transaction id to LogChannel
Catchup wants to be able to know which is the starting transaction id for the log channel.
We use this at the client/writer end to figure out which is the last transaction appended to the file
when doing log rotation (because kernel heads to the transaction log file header).