Fix append log rotation
As it was we would rotate in the middle of a transaction.
If the transaction id is present, that means that we can rotate before appending,
because the buffer starts a new transaction. There is no guarantee that will end a transaction.

Tests changed because the number of rotations decreased by one (the first append being made doesn't trigger a rotation anymore)