Persist new log version on disk before rotation ()
* Persist new log version on disk before rotation

When pulling transactions we need kernel to be able to resume from the latest
log file in case we have multiple pulling "sessions" (or crash in between "sessions") before recovering.

As it was, the batched rotation was not persisting the latest log version in the metadatastore.
This resulted in the following scenario:

- I pulled transactions into log file 0 and 1.
- I stop
- I pulled a few more transactions using a new session (implies creating all of in-memory state from scratch).
  Because my metadatastore does not have the log 1 registered in the log repo, it goes and uses log 0 as the one to append to.

Because log 0 is full, it rotates and picks log 2 to append. This might sound ok, but it definitely isn't. I found two issues in that state:

- The DumpLogicalLog can't bridge from log 1 to log 2 when printing the folder (despite log 2 definitely containing transactions)
- LogFiles#getTailInformation crashes validating the log files on AbstractLogTailScanner#verifyLastFile
(he is not happy that log 1 does not fill up the pre-allocated bytes)